pyproject.toml
[tool.poetry]
name = "back"
version = "0.1.0"
description = ""
authors = ["Semyon Shikota <toylenium@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0"
djangorestframework = "^3.14.0"
drf-spectacular = "^0.26.5"
django-cors-headers = "^4.3.1"
pillow = "^10.1.0"
psycopg2 = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
